norminette a little
